Allow single child card to select parent
If a card in a higher row is covered by only a single "child" card, and the child and parent add up to 13, allow them to be selected/removed together. This is a weird edge case, but is allowed in some variations of Pyramid, including Tut's Tomb (the game this implementation is based on).